Update theme to swap colors, use Paper, and be a little more square
