+++ src/aboutMe/AboutMe.js
@@ -1,119 +1,13 @@
-import React, { useState } from "react";
-import { Colors, Elevation } from "../Constants";
-import AnimatedBar, { barDirections } from "./AnimatedBar";
-import Header from "../Header";
-import { useSpring, animated } from "react-spring";
+import React from "react";
+import useCheckMobileScreen from "../hooks/useCheckMobileScreen";
+import AboutMeDesktop from "./AboutMeDesktop";
+import AboutMeMobile from "./AboutMeMobile";
 
 export default function AboutMe(props) {
-  const [animationLevel, setAnimationLevel] = useState(0)
-  let showBars = animationLevel > 0;
-  let backOpacity = showBars ? 1 : 0;
-  const upAnim = useSpring({
-    from: { marginTop: "50vw" },
-    to: { marginTop: "0" },
-    onRest: () => setAnimationLevel(1),
-  })
+  const isMobile = useCheckMobileScreen();
 
-  return (
-    <div style={styles.container}>
-      <Header endOpacity={backOpacity} style={{ marginRight: "0.5em" }} />
-
-      <AnimatedBar
-        show={showBars}
-        size={5}
-        direction={barDirections.DOWN}
-        addedMargin={41}
-        style={{ marginRight: "-1em" }}
-        sizeUnit={"vw"}
-      />
-      <AnimatedBar
-        show={showBars}
-        size={5.5}
-        direction={barDirections.RIGHT}
-        style={{ marginBottom: "46vw" }}
-        sizeUnit={"vw"}
-      />
-        <animated.div style={{  ...styles.floatingContainer, ...upAnim, }}>
-          <div style={styles.textContainer}>
-            <h1 style={styles.titleText}>Hey!</h1>
-            <AnimatedBar
-              show={showBars}
-              size={40}
-              direction={barDirections.RIGHT}
-              thickness="0.75em"
-              sizeUnit={"%"}
-            />
-
-            <p style={styles.descriptionText}>
-              Big cat fan over here. Also big video game fan. I think AR/VR
-              are the absolute coolest. I currently have a Valve Index, as
-              well as a Google Daydream (RIP), but I've only done a small
-              bit of VR development. I like AR development though! (link to
-              be added to ARTableTop)
-            </p>
-            <p style={styles.descriptionText}>
-              I go by Tavarus, Tavaru5, Kog Mizzle (on League), and Sovellis
-              Nailo, King of Iostros. My cats call me "foodgiver" and
-              "pillow". My wife just calls me Silas.
-            </p>
-          </div>
-          <img style={styles.meImage} src={require("../images/me.jpg")} />
-        </animated.div>
-      <AnimatedBar
-        show={showBars}
-        size={5.5}
-        direction={barDirections.LEFT}
-        style={{ marginTop: "46vw" }}
-        sizeUnit={"vw"}
-      />
-
-      <AnimatedBar
-        show={showBars}
-        size={5}
-        direction={barDirections.UP}
-        addedMargin={41}
-        style={{ marginLeft: "-1em" }}
-        sizeUnit={"vw"}
-      />
-    </div>
-  );
+  if (isMobile) {
+    return <AboutMeMobile />
+  }
+  return <AboutMeDesktop />
 }
-
-const styles = {
-  container: {
-    backgroundColor: Colors.boneWhite,
-    justifyContent: "center",
-    display: "flex",
-    alignItems: "center",
-    minHeight: "100vh",
-  },
-  floatingContainer: {
-    backgroundColor: Colors.oceanGray,
-    display: "flex",
-    width: "60%",
-    minHeight: "36vw",
-    flexWrap: "wrap",
-    boxShadow: Elevation.niceLilElevation,
-    overflow: "auto",
-  },
-  titleText: {
-    fontSize: "2em",
-    color: "black",
-    fontWeight: "bold",
-    marginBottom: "6px",
-  },
-  textContainer: {
-    fontSize: "1.5em",
-    flex: 1,
-    paddingLeft: "2em",
-    paddingRight: "2em",
-  },
-  descriptionText: {
-    color: "black",
-    fontSize: "1em",
-  },
-  meImage: {
-    height: "100%",
-    maxHeight: "36vw",
-  },
-};
